Extract InputParameterContainer::add to own header
