+++ src/core/io/src/4C_io_input_spec_builders.hpp
@@ -10,7 +10,7 @@
 
 #include "4C_config.hpp"
 
-#include "4C_io_input_parameter_container.hpp"
+#include "4C_io_input_parameter_container.templates.hpp"
 #include "4C_io_input_spec.hpp"
 #include "4C_io_value_parser.hpp"
 #include "4C_io_yaml.hpp"
